kcw-파일간 코드 이동
<refactor> : damage 파일 -> damageProcess 파일
#
damage 파일 삭제 후 내부 데이터 이동
