+++ DamageProcess.cs
@@ -7,7 +7,64 @@ using System.Threading.Tasks;
 namespace TeamProject
 {
     public class DamageProcess
-    {        
+    {
+        public static double skillMinDamage;
+        public static double skillMaxDamage;
+
+        public static int SkillEffect(Skill.SkillType type, int skillnumb)
+        {
+            Random rand = new Random();
+
+            if (type == Skill.SkillType.Attack)
+            {
+                skillMinDamage = Math.Round((double)(Player.player.atk + Skill.characterSkill[skillnumb].skillDamage) * 0.9);
+                skillMaxDamage = Math.Round((double)(Player.player.atk + Skill.characterSkill[skillnumb].skillDamage) * 1.1);
+
+                int damage;
+                damage = rand.Next((int)skillMinDamage, (int)skillMaxDamage);
+                return damage;
+            }
+            else if (type == Skill.SkillType.AttackPercent)
+            {
+                skillMinDamage = Math.Round((double)(Player.player.atk * Skill.characterSkill[skillnumb].skillDamage) * 0.9);
+                skillMaxDamage = Math.Round((double)(Player.player.atk * Skill.characterSkill[skillnumb].skillDamage) * 1.1);
+
+                int damage;
+                damage = rand.Next((int)skillMinDamage, (int)skillMaxDamage);
+                return damage;
+            }
+            else if (type == Skill.SkillType.Defense)
+            {
+                Player.player.def += Skill.characterSkill[skillnumb].skillDamage;
+            }
+            else if (type == Skill.SkillType.DefensePercent)
+            {
+                Player.player.def += Player.player.def * Skill.characterSkill[skillnumb].skillDamage / 100;
+            }
+            else if (type == Skill.SkillType.Heal)
+            {
+                Player.player.hp += Skill.characterSkill[skillnumb].skillDamage;
+            }
+            else if (type == Skill.SkillType.HealPercent)
+            {
+                Player.player.hp += Player.player.hp * Skill.characterSkill[skillnumb].skillDamage / 100;
+            }
+            else if (type == Skill.SkillType.Support)
+            {
+
+            }
+            else
+            {
+                if (Skill.characterSkill[skillnumb].skillname == "백전백승")
+                {
+
+                }
+
+            }
+
+            return 0;
+
+        }
         public int Victim(int atk, int hp, out int isDamage, out bool enemyAvoidanceTrue, out bool enemyCriticalTrue)
         {
             isDamage = atk - Player.player.def;
